+++ src/app.features/board/screens/CommentEditScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import Header from 'src/app.components/Header';
 
 interface Props {
